+++ packages/spec_cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - Added support for `--update-goldens` flag (thanks to @trejdych)
 
+## 0.1.4
+
+- feat: Allow running spec_cli with `dart pub global run spec_cli` (#22)
+- upgrade dependencies
+
 ## 0.1.3
 
 Upgrade Riverpod dependency